# openapi.json
{"openapi":"3.1.0","info":{"title":"Clinical Recommendation API","description":"This API provides endpoints for evaluating patient data, generating clinical recommendations, and retrieving stored recommendations.","contact":{"name":"João Costa","email":"joaoassuncaoecosta@gmail.com"},"version":"1.0.0"},"paths":{"/evaluate":{"post":{"summary":"Evaluate Patient","description":"Evaluate patient data and generate a clinical recommendation.\n\nThis endpoint accepts patient data, validates it, and generates a clinical recommendation\nbased on predefined rules or using an AI model. The recommendation is stored in the database\nand published to RabbitMQ.\n\n**Request Body**:\n- `payload`: A dictionary containing the `patient_data` property.\n\n**Example**:\n```json\n{\n    \"patient_data\": {\n        \"name\": \"John Doe\",\n        \"age\": 45,\n        \"height\": 175,\n        \"weight\": 85,\n        \"recent_surgery\": false,\n        \"ai_description\": \"Experiencing mild joint pain\"\n    }\n}\n```\n\n**Response**:\n- Returns a recommendation object containing:\n    - [patient_id](http://_vscodecontentref_/0): Unique ID for the patient.\n    - [recommendation_id](http://_vscodecontentref_/1): Unique ID for the recommendation.\n    - [recommendation](http://_vscodecontentref_/2): The generated recommendation.\n    - [timestamp](http://_vscodecontentref_/3): The timestamp of when the recommendation was created.\n\n**Example**:\n```json\n{\n    \"patient_id\": \"123e4567-e89b-12d3-a456-426614174000\",\n    \"recommendation_id\": \"123e4567-e89b-12d3-a456-426614174001\",\n    \"recommendation\": \"Weight Management Program\",\n    \"timestamp\": \"2025-04-15T12:00:00Z\"\n}\n```\n\n**Errors**:\n- `400 Bad Request`: If [patient_data](http://_vscodecontentref_/4) is missing or invalid.\n- `401 Unauthorized`: If the JWT token is invalid or expired.\n- `500 Internal Server Error`: If there is an issue with RabbitMQ or the database.","operationId":"evaluate_patient_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recommendation"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/recommendation/{recommendation_id}":{"get":{"summary":"Get Recommendation","description":"Retrieve a previously generated recommendation by its ID.\n\nThis endpoint retrieves a recommendation from the Redis cache or the database\nusing the provided [recommendation_id](http://_vscodecontentref_/5). If the recommendation is found, it is returned.\nUnauthorized users cannot access recommendations created by others.\n\n**Path Parameter**:\n- [recommendation_id](http://_vscodecontentref_/6): The unique ID of the recommendation to retrieve.\n\n**Response**:\n- Returns a recommendation object containing:\n    - [patient_id](http://_vscodecontentref_/7): Unique ID for the patient.\n    - [recommendation_id](http://_vscodecontentref_/8): Unique ID for the recommendation.\n    - [recommendation](http://_vscodecontentref_/9): The generated recommendation.\n    - [timestamp](http://_vscodecontentref_/10): The timestamp of when the recommendation was created.\n\n**Example**:\n```json\n{\n    \"patient_id\": \"123e4567-e89b-12d3-a456-426614174000\",\n    \"recommendation_id\": \"123e4567-e89b-12d3-a456-426614174001\",\n    \"recommendation\": \"Weight Management Program\",\n    \"timestamp\": \"2025-04-15T12:00:00Z\"\n}\n```\n\n**Errors**:\n- `400 Bad Request`: If the [recommendation_id](http://_vscodecontentref_/11) is not a valid UUID.\n- `401 Unauthorized`: If the JWT token is invalid or expired.\n- `404 Not Found`: If the recommendation does not exist or the user is unauthorized to access it.\n- `500 Internal Server Error`: If there is an issue retrieving the recommendation.","operationId":"get_recommendation_recommendation__recommendation_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"recommendation_id","in":"path","required":true,"schema":{"type":"string","title":"Recommendation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recommendation"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"Login","description":"Authenticate a user and return a JWT token.\n\nThis endpoint validates the user's credentials (username and password) against the database.\nIf the credentials are valid, a JWT token is generated and returned.\n\n**Request Body**:\n- [data](http://_vscodecontentref_/12): A dictionary containing the user's [username](http://_vscodecontentref_/13) and [password](http://_vscodecontentref_/14).\n\n**Example**:\n```json\n{\n    \"username\": \"my_admin\",\n    \"password\": \"password123\"\n}\n```\n\n**Response**:\n- Returns a JWT token object containing:\n    - [access_token](http://_vscodecontentref_/15): The generated JWT token.\n    - [token_type](http://_vscodecontentref_/16): The type of token (e.g., \"bearer\").\n\n**Example**:\n```json\n{\n    \"access_token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\",\n    \"token_type\": \"bearer\"\n}\n```\n\n**Errors**:\n- `401 Unauthorized`: If the username or password is invalid.","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"Recommendation":{"properties":{"patient_id":{"type":"string","title":"Patient Id"},"recommendation_id":{"type":"string","title":"Recommendation Id"},"recommendation":{"type":"string","title":"Recommendation"},"timestamp":{"type":"string","title":"Timestamp"}},"type":"object","required":["patient_id","recommendation_id","recommendation","timestamp"],"title":"Recommendation","description":"Model for the recommendation response."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"TokenResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}